src/main/java/com/online/library/service/CommentService.java
package com.online.library.service;

import com.online.library.domain.Comment;
import com.online.library.domain.User;
import com.online.library.domain.Views;
import com.online.library.dto.EventType;
import com.online.library.dto.ObjectType;
import com.online.library.repository.CommentRepo;
import com.online.library.util.WsSender;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.function.BiConsumer;

@Service
public class CommentService {
    private final CommentRepo commentRepo;
    private final BiConsumer<EventType, Comment> wsSender;

    @Autowired
    public CommentService(CommentRepo commentRepo, WsSender wsSender)
    {
        this.commentRepo = commentRepo;
        this.wsSender = wsSender.getSender(ObjectType.COMMENT, Views.FullComment.class);
    }

    public Comment create(Comment comment, User user) {
        comment.setAuthor(user);
        Comment commentFromDb = commentRepo.save(comment);

        wsSender.accept(EventType.CREATE, commentFromDb);

        return commentFromDb;
    }
}